validacao de campo hotel voo

+++ agenciaturismo/src/main/java/br/senac/tads3/pi03b/gruposete/models/Hotel.java
@@ -74,7 +74,7 @@ public class Hotel {
         this.quantidade_hospedes = quantidade_hospedes;
     }
 
-    public double getPreco() {
+    public float getPreco() {
         return preco;
     }
